requirements.txt
aiofiles==23.2.1
aiogram==2.25.1
aiohttp==3.8.6
aiosignal==1.3.1
alt-profanity-check==1.3.2
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.1.0
Babel==2.9.1
beautifulsoup4==4.12.2
bleach==6.0.0
certifi==2023.11.17
charset-normalizer==3.3.2
config==0.5.1
frozenlist==1.4.0
idna==3.6
joblib==1.3.2
magic-filter==1.0.12
matplotlib-inline==0.1.6
mistune==2.0.5
multidict==6.0.4
nbconvert==7.4.0
nest-asyncio==1.5.6
numpy==1.26.2
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.38
psutil==5.9.5
pure-eval==0.2.2
pydantic==2.5.2
pydantic_core==2.14.5
pygame==2.5.2
Pygments==2.15.1
python-decouple==3.8
pytz==2023.3.post1
pywin32==306
pyzmq==25.0.2
scikit-learn==1.3.2
scipy==1.11.4
stack-data==0.6.2
threadpoolctl==3.2.0
tornado==6.3.1
traitlets==5.9.0
typing_extensions==4.9.0
wcwidth==0.2.6
yarl==1.9.4
